pyproject.toml
[project]
name = "yonda-local"
version = "0.1.0"
description = "Local infrastructure and integration test harness for Yonda services."
authors = [{ name = "Yonda Engineers" }]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "anyio>=4.3",
    "boto3>=1.35",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pydantic>=2.12.5",
    "python-ulid>=3.1.0",
]
